requirements.txt
fastapi
uvicorn
requests
requests-oauthlib
python-dotenv
